refactor: useTimer
